requirements.txt
python-dotenv==1.0.0
json5==0.9.11
PyPDF2==3.0.0
pdfplumber==0.9.0
openai==1.30.1
faiss-cpu==1.9.0
streamlit==1.25.0
torch==2.5.1
torchvision==0.20.1
transformers==4.46.2
tqdm==4.66.1
requests==2.31.0
loguru==0.7.0
pytest==7.4.0
pytest-mock==3.10.0
spacy==3.8.2